+++ problem_sop.py
