+++ tests/test_problem_sop.py
@@ -1,9 +1,10 @@
-import unittest
 import json
-from itertools import permutations, count, filterfalse
+import unittest
+from itertools import permutations
+
 from jsonschema import ValidationError
 
-from problem import load_val_json
+from problem_sop import load_val_json
 
 
 class TestLoadValJSON(unittest.TestCase):
